added navbar and footer. Tried to add burger for phones.
added a navbar and a footer. Added some javascript so that when the page is decrypted the colours change. Started trying to add a burger for phone screens - added three spans with fixed height and width to look like the burger. Trying to use javascript so that it appears on phones. still some issues