navbar and footer
changed the divs for each section from having a class to an id, and then adding them as links in the navbar. Made the maincontainer 80vh (so that it fit in between the navbar and footer). Made it scrollable so that you can scroll down each section. Made each div (home, about, whycoding etc) 80vh so that they are also each the size of the screen. I noticed there was a problem with my decryption alogirthm (throwing up mistakes when it decrypts the whycoding section) - will have to debug this later.

Used what I had learnt from playing around on flexbox froggy to simplify the footer css