insert image carousel
struggled with this one for quite a while. It was initially overlapping with the #home section, so I added borders around the seperate sections to see what was going on. The I had problems with the dots from the image carousel not displaying below, until I realised I should just add them into the slideshow container div.